+++ src/test/kotlin/eu/europa/ec/eudi/openid4vci/IssuanceNotificationTest.kt
@@ -15,8 +15,8 @@
  */
 package eu.europa.ec.eudi.openid4vci
 
-import eu.europa.ec.eudi.openid4vci.internal.NotificationEventTO
-import eu.europa.ec.eudi.openid4vci.internal.NotificationTO
+import eu.europa.ec.eudi.openid4vci.internal.formats.NotificationEventTO
+import eu.europa.ec.eudi.openid4vci.internal.formats.NotificationTO
 import io.ktor.client.engine.mock.*
 import io.ktor.http.*
 import io.ktor.http.content.*
